+++ tests/testthat/test_rDolphin.R
@@ -9,9 +9,9 @@ dataset <- new_nmr_dataset_1D(ppm_axis = 1:10,
 dataset[["metadata"]][["external"]][["SubjectID"]] = c("10", "20", "30", "40", "50", "60", "70", "80", "90")
 dataset[["metadata"]][["external"]][["Group"]] = c("a", "a", "a", "a", "a", "b", "b", "b", "b")
 
-blood = files_to_rDolphin(dataset, blood)
-cell = files_to_rDolphin(dataset, cell)
-urine = files_to_rDolphin(dataset, urine)
+blood = files_to_rDolphin(dataset, "blood")
+cell = files_to_rDolphin(dataset, "cell")
+urine = files_to_rDolphin(dataset, "urine")
 
 expect_true(is.list(blood))
 expect_true(is.list(cell))